Note that multiple keys should be unnecessary
This has been a point of confusion in the past. When they create a key it's associated with an account and sometimes people (understandably) assume that association means the key won't work on other accounts.

Not sure what the wording should be, but if we decide to note it, here's one possible approach.